Add Stmt and BB parameter to addMemoryAccess and callers. NFC.
This prepares for the case when Instruction and/or BB can be nullptr,
eg. when inserting a scalar access that is mapped to a particular array
element. Such an accesses can be though of to be located at the
beginning or and of a statement, not to a specific BasicBlock or
Instruction.